requirements.txt
fastapi>=0.68.1
uvicorn>=0.15.0
python-multipart>=0.0.5
pytesseract>=0.3.8
Pillow>=9.0.0
opencv-python-headless>=4.5.3
numpy>=1.21.0
python-jose>=3.3.0
pydantic>=1.8.2 